Relocate timeseries metrics modules (#658)
